+++ R/revdep.R
@@ -215,18 +215,30 @@ revdep_check_reset <- function(pkg = ".") {
 }
 
 revdep_check_from_cache <- function(pkg, cache) {
-  # Install this package into revdep library -----------------------------------
+  # Install all dependencies for this package into revdep library --------------
   if (!file.exists(cache$libpath)) {
     dir.create(cache$libpath, recursive = TRUE, showWarnings = FALSE)
   }
   message(
-    "Installing ", pkg$package, " ", pkg$version,
-    " and dependencies to ", cache$libpath
+    "Installing dependencies for ", pkg$package, " to ", cache$libpath
   )
-  withr::with_libpaths(cache$libpath, action = "prefix", {
-    install(pkg, reload = FALSE, quiet = TRUE, dependencies = TRUE)
+
+  withr::with_libpaths(cache$libpath, {
+    install_deps(pkg, reload = FALSE, quiet = TRUE, dependencies = TRUE)
   })
-  on.exit(remove.packages(pkg$package, cache$libpath), add = TRUE)
+
+  # Always install this package into temporary library, to allow parallel ------
+  # revdep checks --------------------------------------------------------------
+  temp_libpath <- tempfile("revdep")
+  dir.create(temp_libpath)
+  on.exit(unlink(temp_libpath, recursive = TRUE), add = TRUE)
+
+  message(
+    "Installing ", pkg$package, " ", pkg$version, " to ", temp_libpath
+  )
+  withr::with_libpaths(c(temp_libpath, cache$libpath), {
+    install(pkg, reload = FALSE, quiet = TRUE, dependencies = FALSE)
+  })
 
   cache$env_vars <- c(
     NOT_CRAN = "false",
@@ -235,6 +247,14 @@ revdep_check_from_cache <- function(pkg, cache) {
     cache$env_vars
   )
   show_env_vars(cache$env_vars)
+
+  # Use combination of temporary path (with own package) and cached libpath
+  # (for everything else) as check path
+  cache$check_libpath <- c(temp_libpath, cache$libpath)
+
+  # Append temporary path to libpath to avoid duplicate installation of this
+  # package
+  cache$libpath <- c(cache$libpath, temp_libpath)
 
   do.call(check_cran, cache)